+++ pkg/frontend/openshiftcluster_putorpatch.go
@@ -22,8 +22,8 @@ import (
 	"github.com/Azure/ARO-RP/pkg/env"
 	"github.com/Azure/ARO-RP/pkg/frontend/middleware"
 	"github.com/Azure/ARO-RP/pkg/operator"
-	"github.com/Azure/ARO-RP/pkg/util/iswimi"
 	"github.com/Azure/ARO-RP/pkg/util/version"
+	wimi "github.com/Azure/ARO-RP/pkg/util/wimi"
 )
 
 var errMissingIdentityParameter error = fmt.Errorf("identity parameter not provided but required for workload identity cluster")
@@ -137,7 +137,7 @@ func (f *frontend) _putOrPatchOpenShiftCluster(ctx context.Context, log *logrus.
 		// Persist identity URL and tenant ID only for managed/workload identity cluster create
 		// We don't support updating cluster managed identity after cluster creation
 		// TODO - use a common function to check if the cluster is a managed/workload identity cluster
-		if iswimi.IsWimi(doc.OpenShiftCluster.Properties) {
+		if wimi.IsWimi(*doc.OpenShiftCluster) {
 			if err := validateIdentityUrl(doc.OpenShiftCluster, putOrPatchClusterParameters.identityURL); err != nil {
 				return nil, err
 			}
